requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.16.1
sqlparse==0.5.5
drf-spectacular==0.29.0
django-filter==25.2
djangorestframework_simplejwt==5.5.1
PyJWT==2.12.1
